refactor register endpoint
